perf(low-bandwidth): code-split ECharts behind a lazy boundary
The Chart element pulls in ECharts (~500KB), which most lessons never use, yet
it was in the initial classroom bundle. Route it through a next/dynamic lazy
boundary (LazyChart) so ECharts is fetched on demand — only when a slide
actually contains a chart. Chart is already client-only (it calls echarts.init
on a DOM ref in an effect), so `ssr: false` matches its existing behaviour.
