+++ components/slide-renderer/components/element/ChartElement/index.tsx
@@ -2,7 +2,7 @@
 
 import type { PPTChartElement } from '@openmaic/dsl';
 import { ElementOutline } from '../ElementOutline';
-import { Chart } from './Chart';
+import { Chart } from './LazyChart';
 
 export { BaseChartElement } from './BaseChartElement';
